### 2019-07-11 #1559 @whedon
[PRE REVIEW]: drms: A Python package for accessing HMI and AIA data
**Submitting author:** @kbg (<a href="http://orcid.org/0000-0002-1361-5712">Kolja Glogowski</a>)
**Repository:** <a href="https://github.com/sunpy/drms" target ="_blank">https://github.com/sunpy/drms</a>
**Version:** v0.5.6
