### 2019-07-11 #1559 @arfon
@kbg this PR should fix your paper: https://github.com/sunpy/drms/pull/29
