### 2019-07-11 f39607a
JOSS paper metadata: Added missing hyphen
