### 2019-07-11 #1559 @xuanxu
👋 @mgckind and @astrom-tom: Would you be willing to be the reviews of this paper? I think your expertise in Python and astrophysics data management makes both of you a great fit for this submission.
