### 2019-07-15 14708a5
JOSS paper: Fixed two typos
