+++ paper/paper.bib
@@ -20,7 +20,7 @@
 	{Malocha}, M. and {Evans}, J. and {Agrawal}, A. and {Leonard}, A.~J. and 
 	{Robitaille}, T.~P. and {Mampaey}, B. and {Campos-Rozo}, J.~I. and 
 	{Kirk}, M.~S.},
-    title = "{SunPy{\mdash}Python for solar physics}",
+    title = "{SunPy -- Python for solar physics}",
   journal = {Computational Science and Discovery},
 archivePrefix = "arXiv",
    eprint = {1505.02563},
