### 2019-07-29 #1559 @mbobra
Hi! The SunPy Project plans to submit [a paper about our v1.0 release](https://github.com/sunpy/sunpy-1.0-paper) and we'd like to cite this paper on `drms` since it is a SunPy affiliated package. Do you know approximately how long it would take to complete this review so we can synchronize our submission?
