### 2019-08-05 #1614 @mgckind
There is one author that contributed to the repo and appears in the Authors.txt (Nabil Freij) which is not on the paper, is that intentionally? just wondering about not missing anybody. Thanks
