### 2019-08-05 sunpy/drms#30 @mgckind
JOSS paper comments
These are comments on the paper itself which is being reviewed [here](https://github.com/openjournals/joss-reviews/issues/1614)

- please add a link to SDO in the intro
- spell out CGI
> webserver hosting the required CGIs.
- add bold text (not in bold) to the quote below ... and **it is** called a *data series*
> is associated with its own database table and called a data series
- it'd be nice to add snippets used to create Figure 1 (at least the `drms` part)

Review link: openjournals/joss-reviews#1614
