### 2019-08-06 sunpy/drms#31 @mgckind
Add contributions and guidelines for support
Since this is not a subpackage of `sunpy` but an affiliated one, it might need its own explicit instructions for how to contribute to the development or how to report issues and code of conduct.
Please add a note in this regards so I can check off the Community guidelines checkbox.

Review link: openjournals/joss-reviews#1614
