### 2019-08-06 sunpy/drms#32 @mgckind
Automated or manual steps for test
I can see the test folder for pytest but it'd be useful to add a way to test installation based on those tests, a few lines in the Readme would be enough.  This would be for the Automated test checkbox

Review link: openjournals/joss-reviews#1614
