### 2019-08-06 #1614 @mgckind
@xuanxu , @kbg. I'm done with the review,
I installed it, looked at the documentation, the paper, I ran a few examples and others and everything seems in order. After addressing the following points this is ready to go:
- I made a comment about the authors above
- I created 3 issues related to the paper, tests and community guidelines, sunpy/drms#30, sunpy/drms#31 and sunpy/drms#32
- I think adding more examples would be useful, as it wasn't very clear how to filter by date or make an end-to-end plot like the one from the paper

Let me know if you have questions and congrats for this complete and useful package
