### 2019-08-07 sunpy/drms#31 @mbobra
Hello @mgckind, thanks for the review!

Here are some responses:

1. The [drms repo](https://github.com/sunpy/drms/community) has already inherited the SunPy Code of Conduct (the yellow dot is a Github bug, but you can look at the Code of Conduct document simply by clicking on it).
2. All the SunPy organization repositories, including drms, are covered by the [SunPy Code of Conduct](https://github.com/sunpy/.github/blob/master/CODE_OF_CONDUCT.md) and the SunPy [contributing document](https://github.com/sunpy/.github/blob/master/CONTRIBUTING.rst).

Is this enough to satisfy the community guidelines?
