### 2019-08-07 #1614 @mgckind
> > There is one author that contributed to the repo and appears in the Authors.txt (Nabil Freij) which is not on the paper, is that intentionally? just wondering about not missing anybody. Thanks
>
> @kbg invited all contributors to coauthor (see [this issue](https://github.com/sunpy/drms/issues/27)) and Nabil declined.

oh I see, yes this is good. Thanks!
