### 2019-08-07 sunpy/drms#31 @mgckind
Yeah, I think it is. It worth mentioning though in the Readme, as it wasn't clear until now. Thanks
