### 2019-08-09 28243da
Added additional license texts
This adds license texts for all individual files that are provided under
a license different from the main license of the package.
