+++ licenses/README.md
@@ -1,0 +1,4 @@
+# Licenses
+
+This directory contains license texts for all individual files that are
+provided under a license different from the main license of the package.
