### 2019-08-09 b653b70
Added Getting Help, Contributing and CoC sections to the README file
