### 2019-08-09 #1614 @kbg
@mgckind : Thanks for the feedback. I'm currently going through the issues you opened. I will leave a comment there, when I think that I have resolved them.

About adding more examples to the online documentation: Have you had a look at the [Tutorial](https://drms.readthedocs.io/en/latest/tutorial.html)? It contains quite some information about how to use the `drms` module.
