### 2019-08-09 8d7f19b
Added section about running tests to the README file
