### 2019-08-09 b22dfb4
JOSS paper: Two smaller text improvements
