+++ paper/paper.md
@@ -40,11 +40,11 @@ The NASA Solar Dynamics Observatory (SDO) spacecraft has continuously observed t
 
 `drms` is a SunPy-affiliated [@SunPy2015] Python package for accessing data hosted by JSOC. A vast majority of these data come from the HMI and AIA instruments, but JSOC also hosts data from the Michelson Doppler Imager [MDI; @Scherrer1995] aboard the Solar and Heliospheric Observatory (SOHO) spacecraft and data from the Interface Region Imaging Spectrometer [IRIS; @DePontieu2014]. The `drms` package allows users to execute complex queries across any number of metadata keywords and export tailored datasets in a variety of formats (including FITS files, movies, and images).
 
-The `drms` package is used as backend for SunPy's JSOC client, but it can also be installed independently from [PyPI](https://pypi.org/project/drms/) using `pip` or from [`conda-forge`](https://anaconda.org/conda-forge/drms) using the `conda` package manager. Source code and documentation are available on [GitHub](https://github.com/sunpy/drms) and the [SunPy website](https://docs.sunpy.org/projects/drms/en/latest) respectively. The `drms` client communicates with DRMS servers using an HTTP/JSON interface. By default the `drms` client connects to the JSOC DRMS, but it can also be configured to access local DRMS installations at other sites, provided the site runs a webserver hosting the required CGIs.
+The `drms` package is used as backend for SunPy's JSOC client, but it can also be installed independently from [PyPI](https://pypi.org/project/drms/) using `pip` or from [`conda-forge`](https://anaconda.org/conda-forge/drms) using the `conda` package manager. Source code and documentation are available on [GitHub](https://github.com/sunpy/drms) and the [SunPy website](https://docs.sunpy.org/projects/drms/en/latest) respectively. The `drms` client communicates with DRMS servers using an HTTP/JSON interface. By default the `drms` client connects to the JSOC DRMS, but it can also be configured to access local DRMS installations at other sites, provided the site runs a webserver hosting the required Common Gateway Interface (CGI) programs.
 
 One advantage of the DRMS is that it decouples metadata from image data. This means that a user can access and export metadata without downloading any image data. Users can also construct queries across any number of metadata keywords, to only download images of interest.
 
-Each data collection in JSOC (for example, AIA images of the solar corona in 171 Å or HMI maps of the photospheric magnetic field in 6173 Å) is associated with its own database table and called a *data series*. An example of a data series with rich, valuable metadata is called `hmi.sharp_720s` and contains Space-weather HMI Active Region Patches [@Bobra14], or SHARPs.
+Each data collection in JSOC (for example, AIA images of the solar corona in 171 Å or HMI maps of the photospheric magnetic field in 6173 Å) is associated with its own database table and it is called a *data series*. An example of a data series with rich, valuable metadata is called `hmi.sharp_720s` and contains Space-weather HMI Active Region Patches [@Bobra14], or SHARPs.
 
 ![Left: Total unsigned flux and mean current helicity of HMI Active Region Patch (HARP) Number 4315 during its disk passage; the dashed orange line marks the central meridian crossing of the active region. Right: Continuum intensity map and magnetic field map at the time of the central meridian crossing.](sharp.pdf)
 
