### 2019-08-09 f881ee7
JOSS paper: Added citation to SDO paper
