+++ paper/paper.md
@@ -36,7 +36,7 @@ bibliography: paper.bib
 
 # Summary
 
-The NASA Solar Dynamics Observatory (SDO) spacecraft has continuously observed the Sun since 2010. It takes about 1.5 terabytes of data per day, in the form of images or spectral data and associated metadata. The data from two of its instruments, Helioseismic and Magnetic Imager [HMI; @schou12] and Atmospheric Imaging Assembly [AIA; @lemen12], are stored and distributed by the Joint Science Operations Center (JSOC) at Stanford University. Specifically, the metadata and pointers to the image data are stored in a PostgreSQL database and managed by the Data Record Management System (DRMS). The data and metadata can be accessed using the [JSOC website](http://jsoc.stanford.edu/ajax/lookdata.html).
+The NASA Solar Dynamics Observatory [SDO; @Pesnell2012] spacecraft has continuously observed the Sun since 2010. It takes about 1.5 terabytes of data per day, in the form of images or spectral data and associated metadata. The data from two of its instruments, Helioseismic and Magnetic Imager [HMI; @schou12] and Atmospheric Imaging Assembly [AIA; @lemen12], are stored and distributed by the Joint Science Operations Center (JSOC) at Stanford University. Specifically, the metadata and pointers to the image data are stored in a PostgreSQL database and managed by the Data Record Management System (DRMS). The data and metadata can be accessed using the [JSOC website](http://jsoc.stanford.edu/ajax/lookdata.html).
 
 `drms` is a SunPy-affiliated [@SunPy2015] Python package for accessing data hosted by JSOC. A vast majority of these data come from the HMI and AIA instruments, but JSOC also hosts data from the Michelson Doppler Imager [MDI; @Scherrer1995] aboard the Solar and Heliospheric Observatory (SOHO) spacecraft and data from the Interface Region Imaging Spectrometer [IRIS; @DePontieu2014]. The `drms` package allows users to execute complex queries across any number of metadata keywords and export tailored datasets in a variety of formats (including FITS files, movies, and images).
 
