+++ paper/paper.bib
@@ -151,3 +151,18 @@ archivePrefix = "arXiv",
    adsurl = {http://adsabs.harvard.edu/abs/2014SoPh..289.2733D},
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@ARTICLE{Pesnell2012,
+   author = {{Pesnell}, W.~D. and {Thompson}, B.~J. and {Chamberlin}, P.~C.
+	},
+    title = "{The Solar Dynamics Observatory (SDO)}",
+  journal = {Solar Physics},
+ keywords = {SDO, Solar cycle, Helioseismology, Coronal, Space weather},
+     year = 2012,
+    month = jan,
+   volume = 275,
+    pages = {3-15},
+      doi = {10.1007/s11207-011-9841-3},
+   adsurl = {https://ui.adsabs.harvard.edu/abs/2012SoPh..275....3P},
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
