### 2019-08-09 efe5276
Updated links to the documentation in README file
