### 2019-08-09 4e401c1
Updated installation instructions
