### 2019-08-09 89e5d5c
Added JOSS paper plot script to examples
