+++ examples/create_sharp_plots.py
@@ -1,0 +1,193 @@
+import os.path
+import numpy as np
+import matplotlib.pyplot as plt
+from matplotlib import dates
+from astropy.io import fits
+import pandas
+import drms
+
+
+def read_fits_data(fname):
+    """Reads FITS data and fixes/ignores any non-standard FITS keywords."""
+    hdulist = fits.open(fname)
+    hdulist.verify('silentfix+warn')
+    return hdulist[1].data
+
+
+do_download = 1
+save_plots = 1
+show_plots = 0
+
+export_email = 'myname@example.com'
+series = 'hmi.sharp_720s'
+sharpnum = 4315
+segments = ['magnetogram', 'continuum']
+# segments = ['magnetogram', 'field', 'continuum', 'Dopplergram']
+fname_fmt_str = '{series}.{sharpnum}.{tstr}.{segment}.fits'
+
+kwlist = ['T_REC', 'LON_FWT', 'OBS_VR', 'CROTA2',
+          'CRPIX1', 'CRPIX2', 'CDELT1', 'CDELT2', 'CRVAL1', 'CRVAL2',
+          'AREA_ACR', 'MEANGAM', 'USFLUX', 'ERRVF', 'MEANJZH', 'ERRMIH']
+
+if do_download:
+    c = drms.Client(email=export_email)
+else:
+    c = drms.Client('kis')
+
+print('Querying metadata...')
+si = c.info(series)
+kw = c.query('%s[%d]' % (series, sharpnum), key=kwlist, rec_index=True)
+t = drms.to_datetime(kw.T_REC)
+
+print('Finding central meridian crossing...')
+rec_cm = kw.LON_FWT.abs().idxmin()
+k_cm = kw.loc[rec_cm]
+t_cm = drms.to_datetime(kw.T_REC[rec_cm])
+print('-> rec_cm:', rec_cm, '@', kw.LON_FWT[rec_cm], 'deg')
+
+# Generate filenames to check if the files are already downloaded.
+t_cm_str = t_cm.strftime('%Y%m%d_%H%M%S_TAI')
+fnames = {
+    s: fname_fmt_str.format(
+        series=series, sharpnum=sharpnum, tstr=t_cm_str, segment=s)
+    for s in segments}
+
+# Check if any files were already downloaded and download them if not.
+download_segments = []
+for k, v in fnames.items():
+    if not os.path.exists(v):
+        download_segments.append(k)
+if do_download and download_segments:
+    print('Downloading files...')
+    exp_query = '%s{%s}' % (rec_cm, ','.join(download_segments))
+    r = c.export(exp_query)
+    dl = r.download('.', verbose=True)
+
+
+print('Reading data files...')
+data = {
+    s: read_fits_data(fnames[s])
+    for s in segments}
+
+# Remove observer's LOS velocity
+if 'Dopplergram' in data:
+    data['Dopplergram'] -= k_cm.OBS_VR
+
+# Convert pixel to world coordinates using WCS keywords
+ny, nx = data[segments[0]].shape
+xmin = (1 - k_cm.CRPIX1)*k_cm.CDELT1 + k_cm.CRVAL1
+xmax = (nx - k_cm.CRPIX1)*k_cm.CDELT1 + k_cm.CRVAL1
+ymin = (1 - k_cm.CRPIX2)*k_cm.CDELT2 + k_cm.CRVAL2
+ymax = (ny - k_cm.CRPIX2)*k_cm.CDELT2 + k_cm.CRVAL2
+
[…]
