### 2019-08-10 eae7b2b
Renamed JOSS plot script, updated code, added more comments
