+++ examples/create_joss_figure.py
@@ -1,0 +1,188 @@
+"""
+This example creates the figure shown in the JOSS paper.
+"""
+from __future__ import absolute_import, division, print_function
+import os.path
+import numpy as np
+import matplotlib.pyplot as plt
+from matplotlib import dates
+from astropy.io import fits
+import example_helpers
+import drms
+
+import pandas
+pandas_version = tuple(map(int, pandas.__version__.split('.')[:2]))
+if pandas_version >= (0, 22):
+    # Since pandas v0.22, we need to explicitely register matplotlib
+    # converters to use pandas.Timestamp objects in plots.
+    pandas.plotting.register_matplotlib_converters()
+
+
+def read_fits_data(fname):
+    """Reads FITS data and fixes/ignores any non-standard FITS keywords."""
+    hdulist = fits.open(fname)
+    hdulist.verify('silentfix+warn')
+    return hdulist[1].data
+
+# Print the doc string of this example.
+print(__doc__)
+
+# This example requires a registered export email address. You can register
+# JSOC exports at: http://jsoc.stanford.edu/ajax/register_email.html
+#
+# You will be asked for your registered email address during execution of
+# this example. If you don't want to enter it every time you run this script,
+# you can set the environment variable JSOC_EXPORT_EMAIL or the variable
+# below to your registered email address.
+email = ''
+
+# Should the plots been shown and/or saved?
+save_plots = True
+show_plots = True
+
+# Series, harpnum and segment selection
+series = 'hmi.sharp_720s'
+sharpnum = 4315
+segments = ['magnetogram', 'continuum']
+fname_fmt_str = '{series}.{sharpnum}.{tstr}.{segment}.fits'
+
+# Download directory
+out_dir = os.path.join('downloads', 'sharp_joss')
+
+# Create download directory if it does not exist yet.
+if not os.path.exists(out_dir):
+    os.makedirs(out_dir)
+
+# Keywords to be queried
+kwlist = ['T_REC', 'LON_FWT', 'CROTA2', 'AREA_ACR', 'USFLUX', 'ERRVF',
+          'CRPIX1', 'CRPIX2', 'CDELT1', 'CDELT2', 'CRVAL1', 'CRVAL2']
+
+# Create DRMS client, use debug=True to see the query URLs.
+c = drms.Client(verbose=True)
+
+print('Querying metadata...')
+kw = c.query('%s[%d]' % (series, sharpnum), key=kwlist, rec_index=True)
+t = drms.to_datetime(kw.T_REC)
+
+print('Finding central meridian crossing...')
+rec_cm = kw.LON_FWT.abs().idxmin()
+k_cm = kw.loc[rec_cm]
+t_cm = drms.to_datetime(kw.T_REC[rec_cm])
+print('-> rec_cm:', rec_cm, '@', kw.LON_FWT[rec_cm], 'deg')
+
+# Check if any files were already downloaded.
+fnames = {}
+download_segments = []
+t_cm_str = t_cm.strftime('%Y%m%d_%H%M%S_TAI')
+for s in segments:
+    fnames[s] = fname_fmt_str.format(
+        series=series, sharpnum=sharpnum, tstr=t_cm_str, segment=s)
+    if not os.path.exists(os.path.join(out_dir, fnames[s])):
+        download_segments.append(s)
+
+# Only download missing files.
[…]
