### 2019-08-10 6c91b2e
Examples: Export email from environment variable
When running the export examples, the user can now also use the environment
variable JSOC_EXPORT_EMAIL to specify their registered export email address.
