### 2019-08-10 c0739e7
Examples: Fixed pandas future warnings
