### 2019-08-12 sunpy/drms#34 @aureliocarnero
Error while running example create_joss_figure.py
Hello, I had an error while running the create_joss_figure example.

Review link: openjournals/joss-reviews#1614

I have tested all the other examples and none gave any error, so it is only in this example.

After running python create_joss_figure.py, the error is the following:

This example creates the figure shown in the JOSS paper.

Querying metadata...
Finding central meridian crossing...
-> rec_cm: hmi.sharp_720s[4315][2014.07.07_05:12:00_TAI] @ -0.038282 deg
Reading image files...
Creating plots...
Traceback (most recent call last):
  File "create_joss_figure.py", line 147, in <module>
    ax_meta[0].set_xticklabels([])
  File "/usr/lib64/python2.7/site-packages/matplotlib/axes/_base.py", line 3295, in set_xticklabels
    minor=minor, **kwargs)
  File "/usr/lib64/python2.7/site-packages/matplotlib/axis.py", line 1653, in set_ticklabels
    ticks = self.get_major_ticks()
  File "/usr/lib64/python2.7/site-packages/matplotlib/axis.py", line 1396, in get_major_ticks
    numticks = len(self.get_major_locator()())
  File "/usr/lib64/python2.7/site-packages/matplotlib/dates.py", line 1249, in __call__
    self.refresh()
  File "/usr/lib64/python2.7/site-packages/matplotlib/dates.py", line 1269, in refresh
    dmin, dmax = self.viewlim_to_dt()
  File "/usr/lib64/python2.7/site-packages/matplotlib/dates.py", line 1026, in viewlim_to_dt
    .format(vmin))
ValueError: view limit minimum -7.02647999992e+16 is less than 1 and is an invalid Matplotlib date value. This often happens if you pass a non-datetime value to an axis that has datetime units

Do you know if it is something real? Or is it maybe my matplotlib version? This is matplotlib: 2.2.2

Cheers
Aurelio
