### 2019-08-12 sunpy/drms#34 @aureliocarnero
I have tried with matplotlib version 2.2.3 and I have the same error.
