### 2019-08-12 sunpy/drms#34 @kbg
Thanks for reporting this issue. It seems be related to the pandas version: I get the same error with "pandas<0.24", but it works fine with "pandas>=0.24" (without changing any other package versions).

I think I have found a workaround for this issue. I'll try to push a fix later.
