### 2019-08-12 2ada0de
Fixed an issue with JOSS plot example and pandas<0.24
There is an issue with pandas < v0.24 when using a pandas.Series() of
pandas.Timestamps() in Matplotlib's plt.errorbar(). To avoid this, we
use the series values for the plt.errorbar() call as a workaround.

This fixes #34.
