### 2019-08-13 sunpy/drms#34 @aureliocarnero
It worked, thanks! Pretty cool code, though. I'm almost finishing with the
validation

El mar., 13 ago. 2019 a las 0:55, Kolja Glogowski (<notifications@github.com>)
escribió:

> It appears that there is an issue with pandas < v0.24 when using
> pandas.Series containing pandas.Timestamps with Matplotlib's errorbar()
> plot.
>
> The following is a minimal example that illustrates this:
>
> import numpy as npimport pandas as pdimport matplotlib.pyplot as plt
>
> pd.plotting.register_matplotlib_converters()
>
> t = pd.Series(pd.date_range('2010-01-01', periods=7, freq='AS'))
> a = pd.Series(np.sin(np.arange(len(t))))
>
> fig, ax = plt.subplots(nrows=2, num=1, clear=True)
> ax[0].plot(t, a)
> ax[1].errorbar(t, a, yerr=0.1)
> plt.show()
>
> This works without problems for "pandas>=0.24", but the errorbar() call
> fails for "pandas<0.24". Curiously, the normal plot() call also works for
> older pandas versions.
>
> To make this work with older pandas versions, one can apparently just use
> the series values instead, when calling errorbar(), i.e.
>
> ax[1].errorbar(t.values, a, yerr=0.1)
>
> which I now also used as a workaround for the JOSS plot example.
>
> I could not find any pandas issues that are directly connected to this,
> but it appears that they have fixed many datetime-related bugs
> <https://pandas.pydata.org/pandas-docs/stable/whatsnew/v0.24.0.html#datetimelike>
> in the v0.24 release, so this appears to be resolved in newer versions.
>
> I closed this issue for now, feel free to re-open it, if you still have
> any problems.
>
