### 2019-08-13 #1614 @aureliocarnero
I have checked all the checklist and in my opinion, it is ready to go.
You have adressed all the issues open by @mgckind and correct the example used to create the publication figure.
Congrats!
