### 2019-08-13 #1614 @xuanxu
@mgckind: There's a couple items still unmarked in your checklist (test + community guidelines) that seems to be addressed,. Can you confirm you completed the review/recommend publication?
