### 2019-08-13 #1614 @kbg
Thanks. I'm also going to add a link to the figure example into the paper. I think that all issues will be resolved after this.

Also, in case the repository branch of the paper source code would show up somewhere in the publication: I will merge the `joss-paper` branch into `master`, after it got accepted.
