### 2019-08-13 a77b5c2
JOSS paper: Added link to figure plot script example
