+++ paper/paper.md
@@ -48,7 +48,7 @@ Each data collection in JSOC (for example, AIA images of the solar corona in 171
 
 ![Left: Total unsigned flux and mean current helicity of HMI Active Region Patch (HARP) Number 4315 during its disk passage; the dashed orange line marks the central meridian crossing of the active region. Right: Continuum intensity map and magnetic field map at the time of the central meridian crossing.](sharp.pdf)
 
-The left two panels of Figure 1 show a small selection of the available metadata which characterize a particular solar active region. From the steep gradient in the total unsigned magnetic flux and the increasing area of active pixels, it is directly evident, without inspecting any image data, that this is a strong emerging active region. The right two panels show a selection of image data corresponding to the time indicated by the dashed orange line in the left two panels. Figure 1 was created from metadata and image data obtained from the JSOC DRMS server using the `drms` package.
+The left two panels of Figure 1 show a small selection of the available metadata which characterize a particular solar active region. From the steep gradient in the total unsigned magnetic flux and the increasing area of active pixels, it is directly evident, without inspecting any image data, that this is a strong emerging active region. The right two panels show a selection of image data corresponding to the time indicated by the dashed orange line in the left two panels. Figure 1 was created from metadata and image data obtained from the JSOC DRMS server using the `drms` package. A Python [script](https://github.com/sunpy/drms/blob/master/examples/create_joss_figure.py) that creates this figure is available in the [examples](https://github.com/sunpy/drms/tree/master/examples) directory of the `drms` source code.
 
 
 # Acknowledgements
