### 2019-08-13 sunpy/drms#30 @kbg
- I added a reference to SDO: f881ee70e38a71050cc19cf3a26b839c5a98d0c1
- I spelled out CGI and added "it is": b22dfb47a46b2dbbe80097504e1da984e905f5a2
- I added the figure plot script to the examples directory and referred to it in the paper: a77b5c2a162d1d3b23c4cb961f6c94a0ded0723b

I talked to @mbobra about adding code snippets to the paper, and we would both rather try to avoid them. Please let me know if you feel that referring to the plot script does not suffice in this regard.
