### 2019-08-13 sunpy/drms#32 @kbg
I added a section about running tests to the README: 8d7f19bf2627f4d696b35ef80e612ca9575f78c6.
