### 2019-08-13 sunpy/drms#30 @mgckind
Yes, I agreed, this is perfect as I looked at the script and a link is just as good. Thanks and congrats!
