### 2019-08-13 #1614 @mgckind
I closed the issues and did a final review, this is good to go. @xuanxu , @kbg . I checked off the pending items. Congratulations
