### 2019-08-13 #1614 @xuanxu
Everything looks good @kbg, here are the next steps:
- please merge the joss-paper branch into master,
- then you should make a new tagged release that includes all changes made during this review
- Archive that release in Zenodo
- Check the Zenodo deposit has the correct metadata: title and author list should match the paper's; you may also add the authors' ORCID.

Once you do that please report here the version number, and the Zenodo DOI.
