### 2019-08-14 #1614 @kbg
Thanks, this is great.

About the author list: Do the author lists (GitHub/Zenodo vs. JOSS paper) really need to be identical? At the moment the GitHub contributors do not match completely with the authors of the paper. We have two GitHub contributors who decided not to be on the author list (see https://github.com/sunpy/drms/issues/27). On the other hand, we also have a paper author who did not directly contribute any code to the GitHub repository, but who is still involved in the project (he is one of the JSOC/DRMS developers).

Please let me know, how I should handle this. I cannot just remove contributors from the Zenodo author list, nor can I put people on the paper author list, without having their consent.
