### 2019-08-14 #1614 @xuanxu
@kbg: The author list of the specific version archived in Zenodo for the paper should match the author list of the paper. No need for any of those to match the Github contributors list, so you should edit just the Zenodo site.
