### 2019-08-14 #1614 @xuanxu
@kbg OK (if you want to use the automatically created by the hook is ok too, the metadata of any Zenodo archive is editable and that includes the author list)
