### 2019-08-16 478957d
Joss paper (#35)
Accepted version of the JOSS paper.

Review: https://github.com/openjournals/joss-reviews/issues/1614
