+++ paper/paper.md
@@ -1,0 +1,58 @@
+---
+title: 'drms: A Python package for accessing HMI and AIA data'
+tags:
+  - Python
+  - astronomy
+authors:
+  - name: Kolja Glogowski
+    orcid: 0000-0002-1361-5712
+    affiliation: "1, 2"
+  - name: Monica G. Bobra
+    orcid: 0000-0002-5662-9604
+    affiliation: 3
+  - name: Nitin Choudhary
+    orcid: 0000-0001-6915-4583
+    affiliation: 4
+  - name: Arthur B. Amezcua
+    orcid: 0000-0002-0361-6463
+    affiliation: 3
+  - name: Stuart J. Mumford
+    orcid: 0000-0003-4217-4642
+    affiliation: 5
+affiliations:
+ - name: Kiepenheuer-Institut für Sonnenphysik, 79104, Freiburg, Germany
+   index: 1
+ - name: eScience Department, Computing Center, University of Freiburg, 79104, Freiburg, Germany
+   index: 2
+ - name: W.W. Hansen Experimental Physics Laboratory, Stanford University, Stanford, CA 94305, USA
+   index: 3
+ - name: Department of Mathematics, Indian Institute of Technology Kharagpur, Kharagpur, West Bengal 721302, India
+   index: 4
+ - name: School of Mathematics and Statistics, The University of Sheffield, Sheffield S3 7RH, UK
+   index: 5
+date: 02 July 2019
+bibliography: paper.bib
+---
+
+# Summary
+
+The NASA Solar Dynamics Observatory [SDO; @Pesnell2012] spacecraft has continuously observed the Sun since 2010. It takes about 1.5 terabytes of data per day, in the form of images or spectral data and associated metadata. The data from two of its instruments, Helioseismic and Magnetic Imager [HMI; @schou12] and Atmospheric Imaging Assembly [AIA; @lemen12], are stored and distributed by the Joint Science Operations Center (JSOC) at Stanford University. Specifically, the metadata and pointers to the image data are stored in a PostgreSQL database and managed by the Data Record Management System (DRMS). The data and metadata can be accessed using the [JSOC website](http://jsoc.stanford.edu/ajax/lookdata.html).
+
+`drms` is a SunPy-affiliated [@SunPy2015] Python package for accessing data hosted by JSOC. A vast majority of these data come from the HMI and AIA instruments, but JSOC also hosts data from the Michelson Doppler Imager [MDI; @Scherrer1995] aboard the Solar and Heliospheric Observatory (SOHO) spacecraft and data from the Interface Region Imaging Spectrometer [IRIS; @DePontieu2014]. The `drms` package allows users to execute complex queries across any number of metadata keywords and export tailored datasets in a variety of formats (including FITS files, movies, and images).
+
+The `drms` package is used as backend for SunPy's JSOC client, but it can also be installed independently from [PyPI](https://pypi.org/project/drms/) using `pip` or from [`conda-forge`](https://anaconda.org/conda-forge/drms) using the `conda` package manager. Source code and documentation are available on [GitHub](https://github.com/sunpy/drms) and the [SunPy website](https://docs.sunpy.org/projects/drms/en/latest) respectively. The `drms` client communicates with DRMS servers using an HTTP/JSON interface. By default the `drms` client connects to the JSOC DRMS, but it can also be configured to access local DRMS installations at other sites, provided the site runs a webserver hosting the required Common Gateway Interface (CGI) programs.
+
+One advantage of the DRMS is that it decouples metadata from image data. This means that a user can access and export metadata without downloading any image data. Users can also construct queries across any number of metadata keywords, to only download images of interest.
+
[…]
