+++ paper/paper.bib
@@ -1,0 +1,168 @@
+@misc{drms,
+  author       = {Kolja Glogowski and
+                  Monica Bobra and
+                  Nabil Freij and
+                  Stuart Mumford and
+                  David Pérez-Suárez and
+                  Nitin Choudhary},
+  title        = {sunpy/drms: drms v0.5.6},
+  month        = feb,
+  year         = 2019,
+  doi          = {10.5281/zenodo.2572850},
+  url          = {https://doi.org/10.5281/zenodo.2572850}
+}
+
+@ARTICLE{SunPy2015,
+   author = {{SunPy Community} and {Mumford}, S.~J. and {Christe}, S. and 
+	{P{\'e}rez-Su{\'a}rez}, D. and {Ireland}, J. and {Shih}, A.~Y. and 
+	{Inglis}, A.~R. and {Liedtke}, S. and {Hewett}, R.~J. and {Mayer}, F. and 
+	{Hughitt}, K. and {Freij}, N. and {Meszaros}, T. and {Bennett}, S.~M. and 
+	{Malocha}, M. and {Evans}, J. and {Agrawal}, A. and {Leonard}, A.~J. and 
+	{Robitaille}, T.~P. and {Mampaey}, B. and {Campos-Rozo}, J.~I. and 
+	{Kirk}, M.~S.},
+    title = "{SunPy -- Python for solar physics}",
+  journal = {Computational Science and Discovery},
+archivePrefix = "arXiv",
+   eprint = {1505.02563},
+ primaryClass = "astro-ph.IM",
+     year = 2015,
+    month = jan,
+   volume = 8,
+   number = 1,
+      eid = {014009},
+    pages = {014009},
+      doi = {10.1088/1749-4699/8/1/014009},
+   adsurl = {http://adsabs.harvard.edu/abs/2015CS%26D....8a4009S},
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{schou12,
+   author = {{Schou}, J. and {Scherrer}, P.~H. and {Bush}, R.~I. and {Wachter}, R. and 
+	{Couvidat}, S. and {Rabello-Soares}, M.~C. and {Bogart}, R.~S. and 
+	{Hoeksema}, J.~T. and {Liu}, Y. and {Duvall}, T.~L. and {Akin}, D.~J. and 
+	{Allard}, B.~A. and {Miles}, J.~W. and {Rairden}, R. and {Shine}, R.~A. and 
+	{Tarbell}, T.~D. and {Title}, A.~M. and {Wolfson}, C.~J. and 
+	{Elmore}, D.~F. and {Norton}, A.~A. and {Tomczyk}, S.},
+    title = "{Design and Ground Calibration of the Helioseismic and Magnetic Imager (HMI) Instrument on the Solar Dynamics Observatory (SDO)}",
+  journal = {Solar Physics},
+ keywords = {Solar Dynamics Observatory, Helioseismology, observations, Instrumentation and data management, Magnetic fields, photosphere},
+     year = 2012,
+    month = jan,
+   volume = 275,
+    pages = {229-259},
+      doi = {10.1007/s11207-011-9842-2},
+   adsurl = {http://adsabs.harvard.edu/abs/2012SoPh..275..229S},
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{lemen12,
+   author = {{Lemen}, J.~R. and {Title}, A.~M. and {Akin}, D.~J. and {Boerner}, P.~F. and 
[…]
