### 2019-08-16 #1614 @mbobra
Hi @xuanxu, I'm writing for @kbg since he is on vacation now. I'm a co-author on the paper and @kbg and I talked about this beforehand -- hopefully this is okay, but if not please let me know.

We are not sure what to do about matching the co-author list on the paper to the author list on the Zenodo release. On one hand, we cannot force people to be authors on a paper if they do not want to. On the other hand, it seems wrong to go into the Zenodo metadata and remove contributors names from the release when they have in fact contributed to the codebase. Therefore, we think it makes sense for the Zenodo release and the paper to have a different list of authors. Please let us know if this is possible.
