### 2019-08-19 #1614 @xuanxu
@mbobra/@kbg: Yes, as long as all the people listed in the paper are also in the Zenodo archive, it is posible to have those different author lists.
