### 2019-08-19 #1614 @mbobra
Thanks, @xuanxu!

- We merged the JOSS paper into the [master branch of the repo](https://github.com/sunpy/drms/tree/master)
- We released a new version of `drms` three days ago, [v0.5.7](https://github.com/sunpy/drms/releases/tag/v0.5.7)
- We archived this release in Zenodo with the DOI [10.5281/zenodo.3369966](https://zenodo.org/record/3369966)
- All the paper authors are also listed in the Zenodo release
