### 2019-08-19 #1614 @xuanxu
@mbobra could you edit the title of the Zenodo archive to:
`drms: A Python package for accessing HMI and AIA data` so it has the same title as the paper?
