### 2019-08-20 #1614 @xuanxu
Yay! :tada:
This is now ready por publication. Pinging @openjournals/joss-eics for final acceptance.
