### 2019-08-20 #1614 @kyleniemeyer
Congrats @kbg on your article's publication in JOSS! Many thanks to @xuanxu for editing,
and @mgckind and @aureliocarnero for reviewing.
