### 2019-08-20 #1614 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
