### 2019-09-03 #1614 @kbg
Thanks for accepting the paper, and also many thanks to everybody for the work they put into editing and reviewing it.

_(I was on vacation, so I'm a bit late to the party :o))._
